test(openapi): merge storeHost upload tests into bohrium test module
Move mock-based OpenAPIContext upload tests into test_run_submission_bohrium
and skip them when bohrium-sdk is not installed, matching CI unittest usage.
